Add Bootstrap Theme to master template.
The express-bootstrap-service package was forked and added to the project.
The package allows simple reference to Bootstrap resources through endpoints
/boostrap/css, /bootstrap/js, /bootstrap/theme, and /bootstrap/fonts.

The forking augmented the original patch to include the Bootstrap Theme.
If the related pull request is accepted by the original author, the project
will switch its dependency to match.